separated fetch in a component, added loading and useEffect, separated date generation in a component, cleared app.jsx
