+++ client/src/styles/categorieitem.css
@@ -1,6 +1,5 @@
 body {
   background-color: black;
-  display: flex;
 }
 
 img {
